Fix some g++-3.3 compilation warnings and errors.
	* mln/debug/histo.hh (h): Disambiguate type.
	* mln/data/sort_psites.hh (to_p_array): Qualify to disambiguate.
	* mln/value/shell.hh (operator): Remove inherited cast operator.
	* mln/value/glf.hh: Upgrade doc style.
	* mln/make/win_chamfer.hh (sqrt): Remove useless cast.
	* tests/data/fill_with_image.cc,
	* tests/data/transform.cc,
	* tests/data/paste.cc (iota): Fix ambiguity; declare then use.
	* tests/data/transform.cc: Avoid sqrt.
	* tests/data/replace.cc: Layout.
	* tests/value/graylevel.cc (test_conversion): Fix ambiguity.
	* tests/morpho/graph_image_wst.cc,
	* tests/morpho/graph_image_morpho.cc,
	* tests/labeling/foreground.cc (mln_VAR): Get rid of it.
	* tests/norm/l2.cc (sqrt): Remove useless return cast.
	Disambiguate argument.


git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@4033 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/mln/make/win_chamfer.hh
@@ -1,5 +1,5 @@
-// Copyright (C) 2006, 2008 EPITA Research and Development Laboratory
-// (LRDE)
+// Copyright (C) 2006, 2008, 2009 EPITA Research and Development
+// Laboratory (LRDE)
 //
 // This file is part of the Milena Library.  This library is free
 // software; you can redistribute it and/or modify it under the terms
@@ -31,7 +31,7 @@
 
 /// \file mln/make/win_chamfer.hh
 ///
-/// Routine to create chamfer mln::w_window2d_int.
+/// \brief Routine to create chamfer mln::w_window2d_int.
 ///
 /// \todo Revamp!
 
@@ -121,7 +121,7 @@ namespace mln
     mln::w_window2d_float
     mk_chamfer_exact()
     {
-      float r2 = static_cast<float>(std::sqrt(2.0f));
+      float r2 = std::sqrt(2.0f);
       float ws[] = { r2, 1, r2,
 		      1, 0,  0,
 		      0, 0,  0 };
